Extract JsonSerializerOptions to reusable static
